Refactor Element, Num, and Analog classes to their own files (no functional changes)
